refactor(auth): move const definition
