+++ src/modules/auth/guard/const.ts
@@ -1,0 +1,6 @@
+export const AUTH_GUARD_PAYLOAD = {
+	USER: "AuthUserPayload",
+	SECURITY_TOKEN: "SecurityTokenPayload",
+	ACCESS_TOKEN: "AccessTokenPayload",
+	TEMP_USER: "TempUser",
+} as const;
