+++ src/modules/quiz/quiz.controller.ts
@@ -35,7 +35,8 @@ import { ArtistService } from "../artist/artist.service";
 import { CheckOwner } from "../auth/metadata/owner";
 import { TokenAuthGuard } from "../auth/guard/authentication/tokenAuth.guard";
 import { OwnerGuard } from "../auth/guard/authorization/owner.guard";
-import { AuthUserPayload, AUTH_GUARD_PAYLOAD } from "../auth/guard/types/requestPayload";
+import { AuthUserPayload } from "../auth/guard/types/requestPayload";
+import { AUTH_GUARD_PAYLOAD } from "../auth/guard/const";
 import { S3Service } from "../aws/s3.service";
 import { DBQueryRunner } from "../db/query-runner/decorator/queryRunner.decorator";
 import { QueryRunnerInterceptor } from "../db/query-runner/queryRunner.interceptor";
